+++ scripts/buildmac.sh
@@ -19,4 +19,9 @@ cc -c ../build/terraformer.c -o ../build/terraformer.o -I/Library/Frameworks/Pyt
 
 cc ../build/terraformer.o -o ../build/terraformer -L/Library/Frameworks/Python.framework/Versions/3.8/lib/python3.8/config-3.8-darwin -lpython3.8 -ldl -framework CoreFoundation
 
+BINDIR="../bin"
+if [ ! -d "$BINDIR" ]; then
+  mkdir ../bin
+fi
+
 cp ../build/terraformer ../bin/terraformer