+++ source/terraformer.py
@@ -23,37 +23,45 @@ import pandas as pd
 
 from messages import *
 from generators import *
+from structures import *
 
 def main():
-   datavars = False
-   individual = False
-   generation = '2'
-   region = 'Dallas'
-   prepend = ''
-   genheader = '# Terraformer generated file'
-   genpath = 'resources'
-   datafolder = 'data'
-   datatype = 'xlsx'
-   propfile = ''
-   propname = '*'
-   propext = 'xlsx'
    print(terraformermessage)
+
    parser = argparse.ArgumentParser(description='Generates Terraform for IBM Virtual Private Cloud')
-   datapath = datafolder
+
+   datapath = useroptions['datapath']
    parser.add_argument('inputvalue', nargs='?', default=datapath, help='input folder or input filename (default: ' + datapath + ')')
+
+   datavars = useroptions['datavars']
+   parser.add_argument('-d', action='store_true', dest='datavars', default=datavars, help='generate variables for data (default: ' + ('True' if datavars else 'False') + ')')
+
+   generation = useroptions['generation']
+   parser.add_argument('-g', dest='generation', default=generation, help='1 for VPC Gen1, 2 for VPC Gen2 (default: ' + generation + ')')
+
+   genpath = useroptions['genpath']
    parser.add_argument('-o', action='store', dest='outputfolder', default=genpath, help='folder for generated Terraform files (default: ' + genpath + ')')
-   parser.add_argument('-g', dest='generation', default='2', help='1 for VPC Gen1, 2 for VPC Gen2 (default: 2)')
-   parser.add_argument('-r', dest='region', default='Dallas', help='region for VPC (default: Dallas)')
-   parser.add_argument('-p', dest='prepend', default='', help='prepend string to resource names (default: \'\')')
-   parser.add_argument('-d', action='store_true', dest='datavars', default=False, help='generate variables for data (default: False)')
+
+   prepend = useroptions['prepend']
+   parser.add_argument('-p', dest='prepend', default=prepend, help='prepend string to resource names (default: ' + prepend + ')')
+
+   puml = useroptions['puml']
+   parser.add_argument('-u', dest='puml', default=puml, help='generate puml (default: ' + ('True' if puml else 'False') + ')')
+
+   region = useroptions['region']
+   parser.add_argument('-r', dest='region', default=region, help='region for VPC (default: ' + region + ')')
+
    parser.add_argument('--version', action='version', version='%(prog)s ' + COPYRIGHT.split(' ')[1])
+
    results = parser.parse_args()
 
-   genpath = results.outputfolder
-   datavars = results.datavars
-   region = results.region
-   prepend = results.prepend
-   generation = results.generation
+   useroptions['datapath'] = results.inputvalue
+   useroptions['datavars'] = results.datavars
+   useroptions['generation'] = results.generation
+   useroptions['genpath'] = results.outputfolder
+   useroptions['prepend'] = results.prepend
+   useroptions['puml'] = results.puml
+   useroptions['region'] = results.region
 
    inputvalue = results.inputvalue
    if (os.path.isdir(inputvalue)):
@@ -68,7 +76,10 @@ def main():
          propext = os.path.splitext(propfilenopath)[1][1:]
          if (os.path.isfile(propfile)):
             found = True
-            generateall(propfile, propname, propext, generation, datavars, genpath, region, individual, prepend)
+            useroptions['propfile'] = propfile
+            useroptions['propname'] = propname
+            useroptions['propext'] = propext
+            generateall(useroptions)
       if (not found):
          print(missinginputmessage % results.inputvalue)
    elif (os.path.isfile(inputvalue)):
@@ -77,14 +88,20 @@ def main():
       propfilenopath = os.path.basename(propfile)
       propname = os.path.splitext(propfilenopath)[0]
       propext = os.path.splitext(propfilenopath)[1][1:]
-      generateall(propfile, propname, propext, generation, datavars, genpath, region, individual, prepend)
+      useroptions['propfile'] = propfile
+      useroptions['propname'] = propname
+      useroptions['propext'] = propext
+      generateall(useroptions)
    elif (os.path.isfile(os.path.join(datapath, inputvalue))):
       # Add path to inputvalue and if found then process the file. 
       propfile = os.path.join(datapath, inputvalue)
       propfilenopath = os.path.basename(propfile)
       propname = os.path.splitext(propfilenopath)[0]
       propext = os.path.splitext(propfilenopath)[1][1:]
-      generateall(propfile, propname, propext, generation, datavars, genpath, region, individual, prepend)
+      useroptions['propfile'] = propfile
+      useroptions['propname'] = propname
+      useroptions['propext'] = propext
+      generateall(useroptions)
    else:
       print(invalidinputmessage % results.inputvalue)